requirements.txt
numpy==1.20.0
scikit-learn==0.24.2
tensorflow-gpu==2.5.0